NETOBSERV-2075: Simplifying observed interfaces - moving it to main map (#509)
* Improve performances by moving observed interfaces to main map

Observed interfaces is a structure managed from the TC hook, so it is
elligible for being written in the main map instead of the additional
metrics map. Moreover, it is very frequently created (the vast majority
of flows being observed from at least two interfaces), so having it in
the additional maps sort of defeats one of its purpose, that was
allowing to reduce the memory used by separating commonly flow data from
more uncommon data.

In other words, when this is set into the addtional
map, the size map becomes similar to main map, whereas if removed the
size map can be shrinked to a much lower number.

lock for observed interface update

Increase max interfaces..

... and ignore direction for duplicate interface checks (tradeoff to
minimize loss of data in case of max reached)

Also some padding optimisation

* Discard non-IP flows

* Include all updates in single lock block

* Manage case with Both directions

format

* Do not report errors on proto/port 0

* Add severity on errors

* Still report errors on 0-ports (e.g. icmp)